+++ packages/calcite-components/src/components/input-number/input-number.tsx
@@ -42,6 +42,7 @@ import {
 import { CSS_UTILITY } from "../../utils/resources";
 import { InputPlacement, NumberNudgeDirection, SetValueOrigin } from "../input/interfaces";
 import { getIconScale } from "../../utils/component";
+import { InternalLabel } from "../functional/InternalLabel";
 import { Validation } from "../functional/Validation";
 import {
   NumericInputComponent,
@@ -63,7 +64,10 @@ declare global {
   }
 }
 
-/** @slot action - A slot for positioning a button next to the component. */
+/**
+ * @slot action - A slot for positioning a `calcite-action` or other interactive content.
+ * @slot label-content - A slot for rendering content next to the component's `labelText`.
+ */
 export class InputNumber
   extends LitElement
   implements
@@ -163,11 +167,11 @@ export class InputNumber
    */
   @property() autocomplete: AutoFill;
 
-  /** When `true`, a clear button is displayed when the component has a value. */
+  /** When present, a clear button is displayed when the component has a value. */
   @property({ reflect: true }) clearable = false;
 
   /**
-   * When `true`, interaction is prevented and the component is displayed with lower opacity.
+   * When present, interaction is prevented and the component is displayed with lower opacity.
    *
    * @mdn [disabled](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/disabled)
    */
@@ -183,7 +187,7 @@ export class InputNumber
    */
   @property({ reflect: true }) form: string;
 
-  /** When `true`, number values are displayed with a group separator corresponding to the language and country format. */
+  /** When present, number values are displayed with a group separator corresponding to the language and country format. */
   @property({ reflect: true }) groupSeparator = false;
 
   /**
@@ -193,16 +197,19 @@ export class InputNumber
    */
   @property({ reflect: true, converter: stringOrBoolean }) icon: IconNameOrString | boolean;
 
-  /** When `true`, the icon will be flipped when the element direction is right-to-left (`"rtl"`). */
+  /** When present, the icon will be flipped when the element direction is right-to-left (`"rtl"`). */
   @property({ reflect: true }) iconFlipRtl = false;
 
-  /** When `true`, restricts the component to integer numbers only and disables exponential notation. */
+  /** When present, restricts the component to integer numbers only and disables exponential notation. */
   @property() integer = false;
 
   /** Accessible name for the component's button or hyperlink. */
   @property() label: string;
 
-  /** When `true`, the component is in the loading state and `calcite-progress` is displayed. */
+  /** When provided, displays label text on the component. */
+  @property() labelText: string;
+
+  /** When present, the component is in the loading state and `calcite-progress` is displayed. */
   @property({ reflect: true }) loading = false;
 
   /**
@@ -275,14 +282,14 @@ export class InputNumber
   @property() prefixText: string;
 
   /**
-   * When `true`, the component's value can be read, but cannot be modified.
+   * When present, the component's value can be read, but cannot be modified.
    *
    * @mdn [readOnly](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/readonly)
    */
   @property({ reflect: true }) readOnly = false;
 
   /**
-   * When `true` and the component resides in a form,
+   * When present and the component resides in a form,
    * the component must have a value in order for the form to submit.
    */
   @property({ reflect: true }) required = false;
@@ -1033,6 +1040,7 @@ export class InputNumber
         placeholder={this.placeholder || ""}
         readOnly={this.readOnly}
         ref={this.childNumberRef}
+        required={this.required}
         type="text"
         value={this.displayedValue}
       />
@@ -1040,6 +1048,14 @@ export class InputNumber
 
     return (
       <InteractiveContainer disabled={this.disabled}>
+        {this.labelText && (
+          <InternalLabel
+            labelText={this.labelText}
+            onClick={this.onLabelClick}
+            required={this.required}
+            tooltipText={this.messages.required}
+          />
+        )}
         <div
           class={{
             [CSS.inputWrapper]: true,