+++ packages/calcite-components/src/components/input/input.tsx
@@ -42,6 +42,7 @@ import {
 } from "../../utils/number";
 import { CSS_UTILITY } from "../../utils/resources";
 import { getIconScale } from "../../utils/component";
+import { InternalLabel } from "../functional/InternalLabel";
 import { Validation } from "../functional/Validation";
 import { IconNameOrString } from "../icon/interfaces";
 import { useT9n } from "../../controllers/useT9n";
@@ -60,7 +61,10 @@ declare global {
   }
 }
 
-/** @slot action - A slot for positioning a `calcite-button` next to the component. */
+/**
+ * @slot action - A slot for positioning a `calcite-button` next to the component.
+ * @slot label-content - A slot for rendering content next to the component's `labelText`.
+ */
 export class Input
   extends LitElement
   implements
@@ -175,11 +179,11 @@ export class Input
    */
   @property() autocomplete: AutoFill;
 
-  /** When `true`, a clear button is displayed when the component has a value. The clear button shows by default for `"search"`, `"time"`, and `"date"` types, and will not display for the `"textarea"` type. */
+  /** When present, a clear button is displayed when the component has a value. The clear button shows by default for `"search"`, `"time"`, and `"date"` types, and will not display for the `"textarea"` type. */
   @property({ reflect: true }) clearable = false;
 
   /**
-   * When `true`, interaction is prevented and the component is displayed with lower opacity.
+   * When present, interaction is prevented and the component is displayed with lower opacity.
    *
    * @mdn [disabled](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/disabled)
    */
@@ -202,23 +206,26 @@ export class Input
    */
   @property({ reflect: true }) form: string;
 
-  /** When `true`, number values are displayed with a group separator corresponding to the language and country format. */
+  /** When present, number values are displayed with a group separator corresponding to the language and country format. */
   @property({ reflect: true }) groupSeparator = false;
 
-  /** When `true`, shows a default recommended icon. Alternatively, pass a Calcite UI Icon name to display a specific icon. */
+  /** When present, shows a default recommended icon. Alternatively, pass a Calcite UI Icon name to display a specific icon. */
   @property({ reflect: true, converter: stringOrBoolean }) icon: IconNameOrString | boolean;
 
-  /** When `true`, the icon will be flipped when the element direction is right-to-left (`"rtl"`). */
+  /** When present, the icon will be flipped when the element direction is right-to-left (`"rtl"`). */
   @property({ reflect: true }) iconFlipRtl = false;
 
   /** Accessible name for the component. */
   @property() label: string;
 
-  /** When `true`, a busy indicator is displayed. */
+  /** When provided, displays label text on the component. */
+  @property() labelText: string;
+
+  /** When present, a busy indicator is displayed. */
   @property({ reflect: true }) loading = false;
 
   /**
-   * When `true`, uses locale formatting for numbers.
+   * When present, uses locale formatting for numbers.
    *
    * @private
    */
@@ -260,7 +267,7 @@ export class Input
   @property({ reflect: true }) minLength: number;
 
   /**
-   * When `true`, the component can accept more than one value.
+   * When present, the component can accept more than one value.
    * This property only has an effect when `type` is "email" or "file".
    * Read the native attribute's documentation on MDN for more info.
    *
@@ -303,14 +310,14 @@ export class Input
   @property() prefixText: string;
 
   /**
-   * When `true`, the component's value can be read, but cannot be modified.
+   * When present, the component's value can be read, but cannot be modified.
    *
    * @mdn [readOnly](https://developer.mozilla.org/en-US/docs/Web/HTML/Attributes/readonly)
    */
   @property({ reflect: true }) readOnly = false;
 
   /**
-   * When `true` and the component resides in a form,
+   * When present and the component resides in a form,
    * the component must have a value in order for the form to submit.
    */
   @property({ reflect: true }) required = false;
@@ -982,6 +989,7 @@ export class Input
         disabled={this.disabled || this.readOnly}
         onClick={this.clearInputValue}
         tabIndex={-1}
+        title={this.messages.clear}
         type="button"
       >
         <calcite-icon icon={ICONS.close} scale={getIconScale(this.scale)} />
@@ -1078,6 +1086,7 @@ export class Input
           placeholder={this.placeholder || ""}
           readOnly={this.readOnly}
           ref={this.childNumberRef}
+          required={this.required}
           type="text"
           value={this.displayedValue}
         />
@@ -1131,6 +1140,15 @@ export class Input
 
     return (
       <InteractiveContainer disabled={this.disabled}>
+        {this.labelText && (
+          <InternalLabel
+            labelText={this.labelText}
+            onClick={this.onLabelClick}
+            required={this.required}
+            tooltipText={this.messages.required}
+          />
+        )}
+
         <div
           class={{
             [CSS.inputWrapper]: true,