+++ packages/calcite-components/src/components/popover/popover.tsx
@@ -113,19 +113,19 @@ export class Popover extends LitElement implements FloatingUIComponent, OpenClos
 
   //#region Public Properties
 
-  /** When `true`, clicking outside of the component automatically closes open `calcite-popover`s. */
+  /** When present, clicking outside of the component automatically closes open `calcite-popover`s. */
   @property({ reflect: true }) autoClose = false;
 
-  /** When `true`, displays a close button within the component. */
+  /** When present, displays a close button within the component. */
   @property({ reflect: true }) closable = false;
 
-  /** When `true`, prevents flipping the component's placement when overlapping its `referenceElement`. */
+  /** When present, prevents flipping the component's placement when overlapping its `referenceElement`. */
   @property({ reflect: true }) flipDisabled = false;
 
   /** Specifies the component's fallback `placement` when it's initial or specified `placement` has insufficient space available. */
   @property() flipPlacements: FlipPlacement[];
 
-  /** When `true`, prevents focus trapping. */
+  /** When present, prevents focus trapping. */
   @property({ reflect: true }) focusTrapDisabled = false;
 
   /**
@@ -165,7 +165,7 @@ export class Popover extends LitElement implements FloatingUIComponent, OpenClos
   /** Offsets the position of the component along the `referenceElement`. */
   @property({ reflect: true }) offsetSkidding = 0;
 
-  /** When `true`, displays and positions the component. */
+  /** When present, displays and positions the component. */
   @property({ reflect: true }) open = false;
 
   /**
@@ -180,7 +180,7 @@ export class Popover extends LitElement implements FloatingUIComponent, OpenClos
   /** Determines where the component will be positioned relative to the `referenceElement`. */
   @property({ reflect: true }) placement: LogicalPlacement = defaultPopoverPlacement;
 
-  /** When `true`, removes the caret pointer. */
+  /** When present, removes the caret pointer. */
   @property({ reflect: true }) pointerDisabled = false;
 
   /**
@@ -200,7 +200,7 @@ export class Popover extends LitElement implements FloatingUIComponent, OpenClos
   @property({ reflect: true }) scale: Scale = "m";
 
   /**
-   * When `true`, disables automatically toggling the component when its `referenceElement` has been triggered.
+   * When present, disables automatically toggling the component when its `referenceElement` has been triggered.
    *
    * This property can be set to `true` to manage when the component is open.
    */