+++ packages/calcite-components/src/custom-theme/input-date-picker.ts
@@ -1,0 +1,51 @@
+import { html } from "../../support/formatting";
+
+export const inputDatePickerTokens = {
+  calciteInputDatePickerBackgroundColor: "",
+  calciteInputDatePickerBorderColor: "",
+  calciteInputDatePickerDividerColor: "",
+  calciteInputDatePickerIconColor: "",
+  calciteInputDatePickerIconColorHover: "",
+  calciteInputDatePickerShadow: "",
+  calciteInputDatePickerInputBackgroundColor: "",
+  calciteInputDatePickerInputBorderColor: "",
+  calciteInputDatePickerInputCornerRadius: "",
+  calciteInputDatePickerInputShadow: "",
+  calciteInputDatePickerInputIconColor: "",
+  calciteInputDatePickerInputIconColorHover: "",
+  calciteInputDatePickerInputTextColor: "",
+  calciteInputDatePickerInputPlaceholderTextColor: "",
+  calciteInputDatePickerDatePickerBorderColor: "",
+  calciteInputDatePickerDatePickerCornerRadius: "",
+  calciteInputDatePickerDatePickerRangeCalendarDividerColor: "",
+  calciteInputDatePickerDatePickerWeekHeaderTextColor: "",
+  calciteInputDatePickerDatePickerHeaderActionBackgroundColor: "",
+  calciteInputDatePickerDatePickerHeaderActionBackgroundColorHover: "",
+  calciteInputDatePickerDatePickerHeaderActionBackgroundColorPress: "",
+  calciteInputDatePickerDatePickerHeaderActionTextColor: "",
+  calciteInputDatePickerDatePickerHeaderActionTextColorPress: "",
+  calciteInputDatePickerDatePickerYearTextColor: "",
+  calciteInputDatePickerDatePickerMonthSelectFontSize: "",
+  calciteInputDatePickerDatePickerMonthSelectTextColor: "",
+  calciteInputDatePickerDatePickerMonthSelectIconColor: "",
+  calciteInputDatePickerDatePickerMonthSelectIconColorHover: "",
+  calciteInputDatePickerDatePickerDayBackgroundColor: "",
+  calciteInputDatePickerDatePickerDayBackgroundColorHover: "",
+  calciteInputDatePickerDatePickerDayTextColor: "",
+  calciteInputDatePickerDatePickerDayTextColorHover: "",
+  calciteInputDatePickerDatePickerCurrentDayTextColor: "",
+  calciteInputDatePickerDatePickerDayBackgroundColorSelected: "",
+  calciteInputDatePickerDatePickerDayTextColorSelected: "",
+  calciteInputDatePickerDatePickerDayRangeTextColor: "",
+  calciteInputDatePickerDatePickerDayRangeBackgroundColor: "",
+  calciteInputDatePickerDatePickerDayOutsideRangeBackgroundColorHover: "",
+  calciteInputDatePickerDatePickerDayOutsideRangeTextColorHover: "",
+};
+
+export const inputDatePicker = html`<calcite-input-date-picker open></calcite-input-date-picker>`;
+export const inputDatePickerRange = html`<calcite-input-date-picker range open></calcite-input-date-picker>`;
+export const inputDatePickerRangeVertical = html`<calcite-input-date-picker
+  range
+  layout="vertical"
+  open
+></calcite-input-date-picker>`;