+++ packages/calcite-components/src/tests/commonTests/index.ts
@@ -12,4 +12,5 @@ export { slots } from "./slots";
 export { labelable } from "./labelable";
 export { themed } from "./themed";
 export { t9n } from "./t9n";
+export { internalLabel } from "./internalLabel";
 export { HYDRATED_ATTR } from "./utils";