+++ packages/calcite-components-react/CHANGELOG.md
@@ -3,6 +3,74 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.0-next.99](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.98...@esri/calcite-components-react@3.3.0-next.99) (2025-09-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.98](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.97...@esri/calcite-components-react@3.3.0-next.98) (2025-09-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.97](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.96...@esri/calcite-components-react@3.3.0-next.97) (2025-09-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.96](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.95...@esri/calcite-components-react@3.3.0-next.96) (2025-09-04)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.95](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.94...@esri/calcite-components-react@3.3.0-next.95) (2025-09-03)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.94](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.93...@esri/calcite-components-react@3.3.0-next.94) (2025-09-03)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.93](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.92...@esri/calcite-components-react@3.3.0-next.93) (2025-09-03)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.92](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.91...@esri/calcite-components-react@3.3.0-next.92) (2025-09-03)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.91](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.90...@esri/calcite-components-react@3.3.0-next.91) (2025-09-02)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.90](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.89...@esri/calcite-components-react@3.3.0-next.90) (2025-09-02)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.89](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.88...@esri/calcite-components-react@3.3.0-next.89) (2025-09-02)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.88](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.87...@esri/calcite-components-react@3.3.0-next.88) (2025-08-29)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.87](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.86...@esri/calcite-components-react@3.3.0-next.87) (2025-08-27)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.86](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.85...@esri/calcite-components-react@3.3.0-next.86) (2025-08-26)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.85](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.84...@esri/calcite-components-react@3.3.0-next.85) (2025-08-26)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.84](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.83...@esri/calcite-components-react@3.3.0-next.84) (2025-08-25)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [3.3.0-next.83](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.82...@esri/calcite-components-react@3.3.0-next.83) (2025-08-22)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [3.3.0-next.82](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@3.3.0-next.81...@esri/calcite-components-react@3.3.0-next.82) (2025-08-21)
 
 **Note:** Version bump only for package @esri/calcite-components-react